deez

+++ src/emulator/emulator.rs
@@ -202,7 +202,7 @@ impl EmulatorState {
                 if $index >= self.program.headers.minheap {
                     self.heap[$index as usize]
                 } else {
-                    self.stack.data[($index as usize - self.program.headers.minheap) as usize]
+                    self.stack.data[($index - self.program.headers.minheap) as usize]
                 }
             };
         }
